+++ cli/sous_manifest_get.go
@@ -8,7 +8,6 @@ import (
 	"github.com/opentable/sous/graph"
 	sous "github.com/opentable/sous/lib"
 	"github.com/opentable/sous/util/cmdr"
-	"github.com/opentable/sous/util/logging"
 	"github.com/opentable/sous/util/yaml"
 	"github.com/pkg/errors"
 )
@@ -18,7 +17,7 @@ type SousManifestGet struct {
 	*sous.ResolveFilter
 	graph.TargetManifestID
 	graph.HTTPClient
-	logging.LogSet
+	graph.LogSink
 	graph.OutWriter
 }