+++ cli/tests/terminal.go
@@ -43,8 +43,8 @@ func NewTerminal(t *testing.T, vstr string) *Terminal {
 	err := io.MultiWriter(baseerr.Buffer, combined.Buffer)
 
 	s := &cli.Sous{Version: v}
-	di := graph.BuildTestGraph(in, out, err)
-	c, er := cli.NewSousCLI(di, s, in, out, err)
+	di := graph.BuildTestGraph(v, in, out, err)
+	c, er := cli.NewSousCLI(di, s, out, err)
 	if er != nil {
 		panic(er)
 	}