+++ ext/docker/image_mapping_test.go
@@ -10,10 +10,10 @@ import (
 	"os/exec"
 	"testing"
 
+	"github.com/nyarly/spies"
 	"github.com/opentable/sous/lib"
 	"github.com/opentable/sous/util/docker_registry"
 	"github.com/opentable/sous/util/logging"
-	"github.com/opentable/sous/util/spies"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"