+++ test/http_state_manager_test.go
@@ -69,7 +69,7 @@ func TestWriteState(t *testing.T) {
 		t.Fatal("State manager double is empty")
 	}
 
-	di := graph.BuildBaseGraph(&bytes.Buffer{}, os.Stderr, os.Stderr)
+	di := graph.BuildBaseGraph(semv.Version{}, &bytes.Buffer{}, os.Stderr, os.Stderr)
 	graph.AddNetwork(di)
 
 	di.Add(