+++ util/docker_registry/dummy_client.go
@@ -3,7 +3,7 @@ package docker_registry
 import (
 	"regexp"
 
-	"github.com/opentable/sous/util/spies"
+	"github.com/nyarly/spies"
 	"github.com/stretchr/testify/mock"
 )